microservices/fetch_data_from_git/developeriq-repository-1/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
awscli==1.31.7
Babel==2.13.1
blinker==1.7.0
boto3==1.33.7
botocore==1.33.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
docutils==0.16
fastapi==0.104.1
Flask==3.0.0
h11==0.14.0
idna==3.6
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
pip==23.2.1
pyasn1==0.5.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
pytimeparse==1.1.8
PyYAML==6.0.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.8.2
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
Werkzeug==3.0.1
zulu==2.0.1

